@jdufresne 2018-01-26
Unfortunately, avoiding `[0]` introduced a [regression](https://code.djangoproject.com/ticket/29067) when `.values_list()` is used with `.annotate()`. I've tried to fix it in PR #9623, but if you have any suggestions regarding performance, please let me know.